+++ src/constants/appVersion.ts
@@ -1,1 +1,1 @@
-export const APP_VERSION = "13.172.10";
+export const APP_VERSION = "13.172.11";